change parent database loader

+++ x_project_adv_worker_db_watcher/models/campaign2categories.py
@@ -1,9 +1,9 @@
-from sqlalchemy import (Column, Integer, ForeignKey)
+from sqlalchemy import (Column, Integer, BigInteger, ForeignKey)
 from .meta import Base
 
 
 class Campaign2Categories(Base):
     __tablename__ = 'campaign2categories'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    id_cam = Column(Integer, ForeignKey('campaign.id'))
+    id_cam = Column(BigInteger, ForeignKey('campaign.id'))
     id_cat = Column(Integer, ForeignKey('categories.id'))
