+++ x_project_adv_worker_db_watcher/models/campaign2device.py
@@ -1,9 +1,9 @@
-from sqlalchemy import (Column, Integer, ForeignKey)
+from sqlalchemy import (Column, Integer, BigInteger, ForeignKey)
 from .meta import Base
 
 
 class Campaign2Device(Base):
     __tablename__ = 'campaign2device'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    id_cam = Column(Integer, ForeignKey('campaign.id'))
+    id_cam = Column(BigInteger, ForeignKey('campaign.id'))
     id_dev = Column(Integer, ForeignKey('device.id'))
