+++ x_project_adv_worker_db_watcher/models/campaign2informer.py
@@ -1,10 +1,10 @@
-from sqlalchemy import (Column, Integer, SmallInteger, ForeignKey)
+from sqlalchemy import (Column, Integer, BigInteger, SmallInteger, ForeignKey)
 from .meta import Base
 
 
 class Campaign2Informer(Base):
     __tablename__ = 'campaign2informer'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    id_cam = Column(Integer, ForeignKey('campaign.id'))
+    id_cam = Column(BigInteger, ForeignKey('campaign.id'))
     id_inf = Column(Integer, ForeignKey('informer.id'))
     allowed = Column(SmallInteger)
