+++ x_project_adv_worker_db_watcher/models/categories2informer.py
@@ -1,4 +1,4 @@
-from sqlalchemy import (Column, Integer, ForeignKey)
+from sqlalchemy import (Column, Integer, BigInteger, ForeignKey)
 from .meta import Base
 
 
@@ -6,4 +6,4 @@ class Categories2Informer(Base):
     __tablename__ = 'categories2informer'
     id = Column(Integer, primary_key=True, autoincrement=True)
     id_cat = Column(Integer, ForeignKey('categories.id'))
-    id_inf = Column(Integer, ForeignKey('informer.id'))
+    id_inf = Column(BigInteger, ForeignKey('informer.id'))
