+++ x_project_adv_worker_db_watcher/models/cron.py
@@ -1,10 +1,10 @@
-from sqlalchemy import (Column, Integer, SmallInteger, Boolean, ForeignKey)
+from sqlalchemy import (Column, BigInteger, SmallInteger, Boolean, ForeignKey)
 from .meta import Base
 
 
 class Cron(Base):
     __tablename__ = 'cron'
-    id_cam = Column(Integer, ForeignKey('campaign.id'), primary_key=True)
+    id_cam = Column(BigInteger, ForeignKey('campaign.id'), primary_key=True)
     day = Column(SmallInteger)
     hour = Column(SmallInteger)
     min = Column(SmallInteger)
