+++ x_project_adv_worker_db_watcher/models/geo.py
@@ -1,9 +1,9 @@
-from sqlalchemy import (Column, Integer, ForeignKey)
+from sqlalchemy import (Column, Integer, BigInteger, ForeignKey)
 from .meta import Base
 
 
 class Geo(Base):
     __tablename__ = 'geo'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    id_cam = Column(Integer, ForeignKey('campaign.id'))
+    id_cam = Column(BigInteger, ForeignKey('campaign.id'))
     id_geo = Column(Integer, ForeignKey('geo_lite_city.id'))
